abstracts into standalone fns for readability
